requirements.txt
torch
transformers
accelerate
sentencepiece
scikit-learn
pandas
numpy
pyyaml
